server/requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
python-multipart==0.0.12
PyJWT==2.9.0
pydantic==2.9.2
scikit-learn==1.5.2
PyPDF2==3.0.1
numpy==2.1.2
scipy==1.13.1
starlette==0.38.5
